+++ docs/changelog.md
@@ -14,6 +14,7 @@ pages](https://github.com/dalibo/ldap2pg/pulls?utf8=%E2%9C%93&q=is%3Apr%20is%3Am
 - Refuse undefined `sync_map`.
 - Fix join when multiple entries are returned.
 - Fix using multiple attributes from joined entries.
+- Fix comment error with generated comment.
 
 
 # ldap2pg 5.2